pyproject.toml
[tool.poetry]
name = "rest-framework"
version = "0.1.0"
description = ""
authors = ["regvud <silveron.regvud@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.0"
isort = "^5.12.0"
django-dotenv = "^1.4.2"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
